hotfix

+++ src/main/java/com/example/ProjektOrliki/team/service/TeamService.java
@@ -8,6 +8,8 @@ import com.example.ProjektOrliki.team.mapper.TeamMapper;
 import com.example.ProjektOrliki.team.model.Team;
 import com.example.ProjektOrliki.team.dto.TeamResponse;
 import com.example.ProjektOrliki.team.repository.TeamRepository;
+import com.example.ProjektOrliki.tournament.model.TournamentStatus;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -61,12 +63,26 @@ public class TeamService {
         return teamMapper.toResponse(team);
     }
 
+    @Transactional
     public void deleteMyTeam() {
         User trainer = currentUserService.getCurrentUser();
 
         Team team = teamRepository.findByTrainer(trainer)
                 .orElseThrow(() -> new IllegalStateException("Nie masz przypisanej drużyny."));
 
+        boolean locked = team.getTournaments().stream()
+                .anyMatch(t -> t.getStatus() != TournamentStatus.CREATED
+                        && t.getStatus() != TournamentStatus.REGISTRATION_OPENED
+                        && t.getStatus() != TournamentStatus.REGISTRATION_CLOSED);
+
+        if (locked) {
+            throw new IllegalArgumentException(
+                    "Nie można usunąć drużyny, ponieważ bierze udział w turnieju, który już wystartował lub został zakończony."
+            );
+        }
+
+        team.getTournaments()
+                .forEach(t -> t.getTeams().remove(team));
         teamRepository.delete(team);
     }
 }
